Allow for later json versions
Do we need to pin the json gem to version 1.8? The test suite seems to
run fine with 2.x versions of json, and having it pinned at 1.8 is
preventing applications that use hydra-role-management from running on
recent systems that come with json 2.x installed by default.

This PR indicates the need for json greater than or equal to 1.8.